tests/e2e/python/pyproject.toml
[tool.poetry]
name = "e2e-python-sdk"
version = "0.1.0"
description = "End-to-end tests for the Goodfire Python SDK"
authors = ["Goodfire"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
anthropic = "^0.30.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.23.8"
pytest-mock = "^3.14.0"
goodfire = "0.3.0.dev.1"
scikit-learn = "^1.6.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"